requirements.txt
cssselect==0.7.1
lxml==4.9.3
leaf==0.4.2
git+https://github.com/dskecse/gae-sessions.git#egg=gaesessions

itsdangerous==1.1.0
click==7.1.2
werkzeug==1.0.1
markupsafe==1.1.1
Jinja2==2.10.1
Flask==1.1.2

six==1.16.0

oauthlib==2.1.0
requests-oauthlib==1.0.0
google-auth-oauthlib==0.4.1

certifi==2021.10.8
pyparsing==2.4.5
packaging==20.9
protobuf==3.17.3
google-api-core==1.31.1

google-auth-httplib2==0.1.0
google-auth==1.19.1

cachetools==3.1.1
requests==2.27.1
httplib2==0.18.1
setuptools==44.1.1
google-api-python-client==1.12.11

requests-toolbelt==0.10.1

gae-env==0.1.1
